refactor: move responses to dto package
