+++ datasource/remote/movie/src/main/java/com/giraffe/movie/dto/MovieByNameResponse.kt
@@ -1,4 +1,4 @@
-package com.giraffe.movie
+package com.giraffe.movie.dto
 
 import com.giraffe.movies.entity.Movie
 import kotlinx.datetime.LocalDate
